fix: rename services
